+++ SheeeeeeeeetExample/Options/FoodOption+ActionSheetItems.swift
@@ -20,7 +20,6 @@ extension FoodOption {
     func item() -> ActionSheetItem {
         return ActionSheetItem(
             title: displayName,
-            subtitle: "oueha",
             value: self,
             image: image)
     }