+++ SheeeeeeeeetTests/Items/ActionSheetItemTests.swift
@@ -8,7 +8,7 @@
 
 import Quick
 import Nimble
-import Sheeeeeeeeet
+@testable import Sheeeeeeeeet
 
 class ActionSheetItemTests: QuickSpec {